backend/requirements_enhanced.txt
# Enhanced requirements for Priority 0-2 improvements
fastapi==0.111.0
uvicorn[standard]==0.30.1
fastmcp==2.12.4
neo4j==5.20.0
pydantic==2.12.0
pydantic-settings==2.11.0
jinja2==3.1.4
httpx==0.28.1
slowapi==0.1.9
limits==3.11.0
python-dotenv==1.1.1
openai==1.35.7
structlog==24.1.0
pyjwt==2.9.0
pytest==8.3.1
pytest-asyncio==0.23.7
pytest-cov==5.0.0
mypy==1.10.0
ruff==0.5.5
coverage==7.5.3
psutil==6.0.0
tenacity==8.2.3

# P0.3: Query Caching
redis==5.0.1

# P1.1: OpenTelemetry Tracing
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-jaeger-thrift==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-neo4j==0.42b0
opentelemetry-instrumentation-requests==0.42b0
